+++ scripts/sd_peft.py
@@ -388,7 +388,7 @@ def train(cfg: SDPeftConfig, strategies: "SdPeftStrategy"):
     train_dataset_group.set_max_train_steps(cfg.training.max_train_steps)
 
     # lr schedulerを用意する
-    lr_scheduler = get_scheduler_fix(cfg.optimizer.scheduler, cfg.optimizer.optimizer_type, cfg.training, optimizer, accelerator.num_processes)
+    lr_scheduler = get_scheduler_fix(cfg.optimizer.scheduler, cfg.optimizer, cfg.training, optimizer, accelerator.num_processes)
 
     # 実験的機能：勾配も含めたfp16/bf16学習を行う　モデル全体をfp16/bf16にする
     if cfg.performance.precision.full_fp16: